feat: agregar registros
